Add files via upload
This is a business plan to a Company that will recycle metal and Sell as product, We will make different materials to constructions and Metallurgic companies.  This company is to be created in Madeira Island, Portugal. The Project is highly Profitable and requires the know-how to work with scrap in different ways. We will look also to be in the trade of recycling Metals On the London Metal Exchange.